VARO-41: removed unnecessary AbstractState
